+++ models/refund.js
@@ -11,7 +11,7 @@ const Refund = {
   req: DataTypes.STRING,
   who: DataTypes.STRING,
   note: DataTypes.STRING,
-  amount: DataTypes.INTEGER,
-  available: DataTypes.INTEGER,
+  amount: DataTypes.FLOAT,
+  available: DataTypes.FLOAT,
 };
 module.exports = Refund;
